requirements.txt
click==8.1.3
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
SQLAlchemy==2.0.8
typing_extensions==4.5.0
Werkzeug==2.2.3
